Return data from database now as LiveData
The actual initial load of the database is now done as a Kotlin
coroutine and querying the database returns all data as LiveData which
is then transformed as needed so no background thread needs to be
directly created.

Robolectric tests are now improved to use actual assets.

Added Robolectric tests for the Summit repository.

Some tests are now currently broken due to the move to LiveData, but the
app itself is working.