Restructured summit list to be a 3-level set of lists for finding summits
This is currently incomplete and left in a broken test for the app, but
it is a large refactor so committing what's currently working. Some of
the existing tests are failing since you can't currently navigate all
the way to summit details.

The top-level fragment is now a list of all the associations which will
then load a fragment with a list of all regions in that association
replacing the spinners originally used. That list will then link into
specific summits once completed.