Summit data downloaded from JSON API only updates appropriate fields
Both the GET region and GET summit download a subset of the columns
available for describing each summit and neither contains all the
information. For example, the locator, valid, and restriction data is
only available through the GET summit JSON data and the last activation
information is only available though the JSON API via the GET region
data structure. Also, both are missing the bonus points for a summit
which is available through the summits CSV file.

With this, no data should be lost or overwritten as it's refreshed from
the various CSV and JSON APIs. All local unit tests are once again
passing.