Remove unrecognized associations on CSV download
This fixes a few more tests if real live data has previously made it
into the app. The test fixture data is missing some of the newer
associations which were becoming stale data when running instrumentation
tests.

This should probably be changed to restict to downloaded JSON APIs for
efficiency.